+++ app/(studio)/studio/[[...index]]/page.tsx
@@ -1,0 +1,10 @@
+"use client";
+
+import { NextStudio } from "next-sanity/studio";
+import config from "../../../../sanity.config";
+
+const StudioPage = () => {
+    return <NextStudio config={config} />;
+};
+
+export default StudioPage;
