+++ app/(studio)/studio/layout.tsx
@@ -1,24 +1,23 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
-import "./globals.scss";
 
 const inter = Inter({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
-    title: "CleanTok",
+    title: "CleanTok Sanity Studio",
     description: "Created by @CleanCode.ro",
 };
 
-const RootLayout = ({
+const Layout = ({
     children,
 }: Readonly<{
     children: React.ReactNode;
 }>) => {
     return (
         <html lang="en">
-            <body className={inter.className}>{children}</body>
+            <body className={`${inter.className} antialiased`}>{children}</body>
         </html>
     );
 };
 
-export default RootLayout;
+export default Layout;
