--- src/pages/studio/[[...index]].tsx
@@ -1,19 +1,0 @@
-import Head from "next/head";
-import { NextStudio } from "next-sanity/studio";
-import { metadata } from "next-sanity/studio/metadata";
-import config from "../../../sanity.config";
-
-const StudioPage = () => {
-    return (
-        <>
-            <Head>
-                {Object.entries(metadata).map(([key, value]) => (
-                    <meta key={key} name={key} content={value} />
-                ))}
-            </Head>
-            <NextStudio config={config} />
-        </>
-    );
-};
-
-export default StudioPage;
